feat(core): move tmux plugin into core (closes #1315)
Relocate src/commands/plugins/tmux/ → src/commands/core/tmux/ (5 files,
~1613 LOC). Wire tmux into CORE_ROUTES + route-tools.ts dispatch. Update
14 cousin imports from sibling plugins, 6 test files (18 refs), and
top-aliases.ts. Delete plugin.json to prevent stale bootstrap symlink.
Re-check routeTools after alias splice so open/close/stall still resolve.

ADR 0001 Q#3 marked Resolved; plugin-audit.md tmux row → core.
