+++ src/commands/core/tmux/safety.ts
