+++ src/commands/core/tmux/stall-detect.ts
