+++ test/isolated/tmux-peek.test.ts
@@ -35,26 +35,26 @@ afterEach(() => {
 
 describe("resolveTmuxTarget — target resolution", () => {
   test("pane ID literal is returned as-is", async () => {
-    const { resolveTmuxTarget } = await import("../../src/commands/plugins/tmux/impl");
+    const { resolveTmuxTarget } = await import("../../src/commands/core/tmux/impl");
     const hit = resolveTmuxTarget("%776");
     expect(hit).toEqual({ resolved: "%776", source: "pane-id" });
   });
 
   test("session:w.p format is passed through", async () => {
-    const { resolveTmuxTarget } = await import("../../src/commands/plugins/tmux/impl");
+    const { resolveTmuxTarget } = await import("../../src/commands/core/tmux/impl");
     const hit = resolveTmuxTarget("101-mawjs:0.1");
     expect(hit).toEqual({ resolved: "101-mawjs:0.1", source: "session:w.p" });
   });
 
   test("team-agent name resolves via team config walk (Bug D fix)", async () => {
-    const { resolveTmuxTarget } = await import("../../src/commands/plugins/tmux/impl");
+    const { resolveTmuxTarget } = await import("../../src/commands/core/tmux/impl");
     const hit = resolveTmuxTarget("known-agent");
     expect(hit?.resolved).toBe("%999");
     expect(hit?.source).toContain("team-agent");
   });
 
   test("team-agent with empty tmuxPaneId falls through to session-name fallback", async () => {
-    const { resolveTmuxTarget } = await import("../../src/commands/plugins/tmux/impl");
+    const { resolveTmuxTarget } = await import("../../src/commands/core/tmux/impl");
     const hit = resolveTmuxTarget("orphan-agent");
     // orphan-agent has tmuxPaneId="" — skipped as not-live, falls through.
     // Per #1012 (don't hardcode :0), the bare name is returned as-is so tmux
@@ -66,7 +66,7 @@ describe("resolveTmuxTarget — target resolution", () => {
   });
 
   test("bare session name → resolved via fleet-stem OR session-name fallback", async () => {
-    const { resolveTmuxTarget } = await import("../../src/commands/plugins/tmux/impl");
+    const { resolveTmuxTarget } = await import("../../src/commands/core/tmux/impl");
     const hit = resolveTmuxTarget("112-fusion");
     expect(hit).not.toBeNull();
     // Per #1012 (don't hardcode :0): resolver returns the bare session name
@@ -77,14 +77,14 @@ describe("resolveTmuxTarget — target resolution", () => {
   });
 
   test("target resolution is deterministic — same input, same output", async () => {
-    const { resolveTmuxTarget } = await import("../../src/commands/plugins/tmux/impl");
+    const { resolveTmuxTarget } = await import("../../src/commands/core/tmux/impl");
     const a = resolveTmuxTarget("known-agent");
     const b = resolveTmuxTarget("known-agent");
     expect(a).toEqual(b);
   });
 
   test("unknown name that looks like session produces fallback (no false-positive match)", async () => {
-    const { resolveTmuxTarget } = await import("../../src/commands/plugins/tmux/impl");
+    const { resolveTmuxTarget } = await import("../../src/commands/core/tmux/impl");
     // Not a team-agent name (not in any config), not a pane-id pattern — fallback to session.
     // Note: may still hit fleet-stem tier if "zzz-nonexistent" word-matches a fleet name.
     // This test isn't isolated from the real fleet dir; just verify we get SOME resolution.
@@ -103,7 +103,7 @@ describe("resolveTmuxTarget — target resolution", () => {
 // resolution before falling through. We verify the source label.)
 describe("resolveTmuxTarget — fleet stem tier (#394 Bug I)", () => {
   test("bare name with no match still produces a resolution with source label", async () => {
-    const { resolveTmuxTarget } = await import("../../src/commands/plugins/tmux/impl");
+    const { resolveTmuxTarget } = await import("../../src/commands/core/tmux/impl");
     // This name won't match any fleet session but WILL match the final fallback.
     const hit = resolveTmuxTarget("definitely-not-a-real-fleet-oracle-xyzzy");
     expect(hit).not.toBeNull();
@@ -115,7 +115,7 @@ describe("resolveTmuxTarget — fleet stem tier (#394 Bug I)", () => {
   });
 
   test("source label for bare-name resolution mentions the tier used", async () => {
-    const { resolveTmuxTarget } = await import("../../src/commands/plugins/tmux/impl");
+    const { resolveTmuxTarget } = await import("../../src/commands/core/tmux/impl");
     const hit = resolveTmuxTarget("some-fleet-candidate-name");
     expect(hit).not.toBeNull();
     // Must be descriptive, not empty
